# [Enhancement] Optimize the display format of the Parquet profile to help showing the relationship of different parts.