## Why I'm doing:
The origin parquet section in profile contains much time information. However, they are not well indented, which make it confusing when analyze the time cost.

## What I'm doing:
1. Optimize the display format of the parquet section in the profile.
2. The origin `GroupChunkReadTime` does not contain the lazy skip time, which make the `GroupChunkReadTime` less than `PageReadTime` when skipping much pages, it is not reasonable and we fix it.

Now, the parquet section in profile looks like: 
```
- Parquet:
	- ParquetReaderInit:
		- InitFooterGetTime: 1m27s
			- ReaderInitFooterRead: 1m19s
			- FooterCacheReadTime: 1.900ms
			- FooterCacheReadCount: 126.101K (126101)
			- FooterCacheWriteTime: 49.523ms
			- FooterCacheWriteBytes: 299.667 GB
			- FooterCacheWriteCount: 24.953K (24953)
			- FooterCacheWriteFailCount: 1.151M (1150803)
		- InitColumnNamesSetTime: 7.396ms
		- InitReadColumnsPrepareTime: 3.531ms
		- InitGroupReaderInitTime: 3m31s
			- ReaderInitColumnReaderInit: 390.094ms
	- ParquetGroupRead:
		- GroupChunkReadTime: 7s107ms
			- PageReadTime: 6s995ms
				- PageDecompressTime: 550.046ms
			- PageReadBytes: 217.216 GB
			- PageReadBytesUncompressed: 294.290 GB
			- PageReadCount: 26.333M (26332985)
			- PageSkipCounter: 43.519K (43519)
			- HasPageStatistics: 503.932K (503932)
			- LevelDecodeTime: 14.077ms
			- ValueDecodeTime: 66.953ms
		- GroupDictFilter: 4.150ms
		- GroupDictDecode: 301.470us
	- IcebergV2FormatTime: 
	  - DeleteFileBuildFilterTime: 0ns
	  - DeleteFileBuildTime: 0ns
	  - DeleteFilesPerScan: 0
	- RequestBytesRead: 2.318 TB
	- RequestBytesReadUncompressed: 2.393 TB
- ReaderInit: 4m58s
- RawRowsRead: 32.948B (32947687172)
- RowsRead: 88.198K (88198)
- ScanRanges: 1.302M (1301857)
```

## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [ ] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [x] Yes, this PR will result in a change in behavior.
- [ ] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with different default values
- [ ] Policy changes: use new policy to replace old one, functionality automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [ ] I have checked the version labels which the pr will be auto-backported to the target branch
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5