@DorianZheng 2024-02-06
And please consider this too
![img_v3_027q_af354d80-49aa-4cbe-998d-9d1f61ff40eg](https://github.com/StarRocks/starrocks/assets/8065637/a060592b-bd47-4873-b91f-feca8565e1bc)