@DorianZheng 2024-02-06
And this 

<img width="368" alt="image" src="https://github.com/StarRocks/starrocks/assets/8065637/3bcc3722-f8f2-4783-8209-1b42c7221a82">